Clip lines.

+++ draw.h
@@ -8,7 +8,7 @@ struct canvas
 };
 
 void
-draw_vline(struct canvas* canvas, int x, int y0, int y1, uint32_t color);
+draw_vline(struct canvas* canvas, size_t x, size_t y0, size_t y1, uint32_t color);
 
 void
 draw_line(struct canvas* canvas, size_t x0, size_t y0, size_t x1, size_t y1, uint32_t color);
